feat: 예약 생성 API 구현

+++ src/main/java/com/company/groomingzone/barbershop/service/dto/CreateBarberShopCommand.java
@@ -1,16 +1,18 @@
 package com.company.groomingzone.barbershop.service.dto;
 
+import java.math.BigDecimal;
+
 public record CreateBarberShopCommand(
         Long ownerId,
         String name,
         String address,
         String phone,
         String introduction,
-        Double latitude,
-        Double longitude
+        BigDecimal latitude,
+        BigDecimal longitude
 ) {
 
-    public static CreateBarberShopCommand of (Long ownerId, String name, String address, String phone, String introduction, Double latitude, Double longitude) {
+    public static CreateBarberShopCommand of (Long ownerId, String name, String address, String phone, String introduction, BigDecimal latitude, BigDecimal longitude) {
         return new CreateBarberShopCommand(ownerId, name, address, phone, introduction, latitude, longitude);
     }
 }
